make operator new handling more generic
Previously, on some specific architectures, an inline placement new was
declared, and on others `<new.h>` was included. But `<new.h>` is a
non-portable header available only in some Arduino architectures, while
declaring our own placement new might conflict with standard libraries
that also declare an inline version (such as uclibc++/ArduinoSTL).

This commit changes the behaviour to always include `<new>` if it is
available (assuming it is complete or at least has placement new, which
seems to be the case for existing Arduino cores). If `<new>` is not
available (which only seems to be the case for megaavr and older avr
cores currently), just declare placement new inline. This has a small
chance of conflict, but seems to be ok for avr and megaavr now.