improve decisions about what (not) to define
Previously, on the (mega)AVR architectures, no standard c++ library was
assumed to be available. Only on these architectures, controlled by the
`ARX_TYPE_TRAITS_DISABLED` macro, all the C++98 and C++11 stuff would be
defined. On all architectures (include (mega)AVR), the `__cplusplus`
macro would be used to decide whether or not to define the C++14 and
C++17 stuff.

This is now made a lot more generic, where `TypeTraits.h` now defines a
single macro with the C++ version supported by the standard library.

In the other headers, this macro is used to decide what to define and
what to include from the standard library. The C++98 and C++11 stuff is
now split, so on e.g. AVR with ArduinoSTL the C++98 stuff can be
included but the C++11 stuff still defined (which prevents defining
stuff that is not needed).

Determining the supported C++ version no longer relies on the Arduino
architecture macros, but instead sees which library is available
exactly. For gcc's libstdc++ and clang's libc++, the compiler
`__cplusplus` version is assumed to be correct, for uclibc++ C++98 is
assumed. Other standard libraries raise an error, since we have no idea
about their capabilities. If no standard library is available (e.g. AVR
without ArduinoSTL, the version macro is set to 0).