move almost everything to arx::arx_std
This declares new stuff in another namespace to prevent conflicts when
standard system libraries are available even when we thought they were
not.

By importing the new arx::arx_std namespace *into* the std namespace,
users of this library do not need to be updated, they can still
reference using the std namespace as before (except that in case of a
conflict, they get the standard library version rather than a
compilation error).

In addition, this imports all of std into arx::arx_std as well, so you
can *also* explicitly reference namespace arx::arx_std for everything,
except that in case of a conflict, you get the arx version rather than
the standard library version (this is relevant when arx defines a newer
version of some name than the standard libraries, such as the C++11 arx
version of numeric_limits vs the C++98 ArduinoSTL version of it).

Note that it would have been nicer to name this namespace arx::std, but
when doing `using namespace arx`, which imports the `std` name, this
creates a conflict with the global `std` namespace.