check __has_include(<initializer_list>)
This ensures that initializer_list is really only defined when not
already available from the standard-named header.

Unlike the other things defined by this library, initializer_list *must*
be defined in std, importing it into std is not enough. This means that
if it would already have been defined elsewhere, this will produce a
conflict (unlike for the other things defined).